hotfix: sonarqube issue 해결
