+++ api/src/main/java/shop/ink3/api/book/book/service/BookService.java
@@ -30,12 +30,12 @@ import shop.ink3.api.book.book.exception.DuplicateIsbnException;
 import shop.ink3.api.book.book.exception.InvalidCategorySelectionException;
 import shop.ink3.api.book.book.external.aladin.dto.AladinBookResponse;
 import shop.ink3.api.book.book.repository.BookRepository;
-import shop.ink3.api.book.bookAuthor.entity.BookAuthor;
-import shop.ink3.api.book.bookAuthor.repository.BookAuthorRepository;
-import shop.ink3.api.book.bookCategory.entity.BookCategory;
-import shop.ink3.api.book.bookCategory.repository.BookCategoryRepository;
-import shop.ink3.api.book.bookTag.entity.BookTag;
-import shop.ink3.api.book.bookTag.repository.BookTagRepository;
+import shop.ink3.api.book.bookauthor.entity.BookAuthor;
+import shop.ink3.api.book.bookauthor.repository.BookAuthorRepository;
+import shop.ink3.api.book.bookcategory.entity.BookCategory;
+import shop.ink3.api.book.bookcategory.repository.BookCategoryRepository;
+import shop.ink3.api.book.booktag.entity.BookTag;
+import shop.ink3.api.book.booktag.repository.BookTagRepository;
 import shop.ink3.api.book.category.dto.CategoryFlatDto;
 import shop.ink3.api.book.category.entity.Category;
 import shop.ink3.api.book.category.exception.CategoryNotFoundException;
