+++ api/src/main/java/shop/ink3/api/book/bookauthor/repository/BookAuthorRepository.java
@@ -1,9 +1,9 @@
-package shop.ink3.api.book.bookAuthor.repository;
+package shop.ink3.api.book.bookauthor.repository;
 
 import java.util.List;
 import org.springframework.data.jpa.repository.EntityGraph;
 import org.springframework.data.repository.CrudRepository;
-import shop.ink3.api.book.bookAuthor.entity.BookAuthor;
+import shop.ink3.api.book.bookauthor.entity.BookAuthor;
 
 public interface BookAuthorRepository extends CrudRepository<BookAuthor, Long> {
     @EntityGraph(attributePaths = "author")
