+++ api/src/main/java/shop/ink3/api/book/bookcategory/repository/BookCategoryRepository.java
@@ -1,10 +1,10 @@
-package shop.ink3.api.book.bookCategory.repository;
+package shop.ink3.api.book.bookcategory.repository;
 
 import java.util.List;
 import org.springframework.data.jpa.repository.EntityGraph;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
-import shop.ink3.api.book.bookCategory.entity.BookCategory;
+import shop.ink3.api.book.bookcategory.entity.BookCategory;
 
 @Repository
 public interface BookCategoryRepository extends JpaRepository<BookCategory, Long> {
