+++ api/src/main/java/shop/ink3/api/book/booktag/entity/BookTag.java
@@ -1,6 +1,5 @@
-package shop.ink3.api.book.bookAuthor.entity;
+package shop.ink3.api.book.booktag.entity;
 
-import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
 import jakarta.persistence.FetchType;
 import jakarta.persistence.GeneratedValue;
@@ -11,22 +10,23 @@ import jakarta.persistence.ManyToOne;
 import jakarta.persistence.Table;
 import lombok.AccessLevel;
 import lombok.AllArgsConstructor;
+import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
-import shop.ink3.api.book.author.entity.Author;
 import shop.ink3.api.book.book.entity.Book;
+import shop.ink3.api.book.tag.entity.Tag;
 
 @Entity
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
 @AllArgsConstructor
+@Builder
 @Getter
-@Table(name = "book_authors")
-public class BookAuthor {
+@Table(name = "book_tags")
+public class BookTag {
 
-    public BookAuthor(Book book, Author author, String role) {
-        this.author = author;
+    public BookTag(Book book, Tag tag) {
         this.book = book;
-        this.role = role;
+        this.tag = tag;
     }
 
     @Id
@@ -38,9 +38,6 @@ public class BookAuthor {
     private Book book;
 
     @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "author_id", nullable = false)
-    private Author author;
-
-    @Column(nullable = false)
-    private String role;
+    @JoinColumn(name = "tag_id", nullable = false)
+    private Tag tag;
 }
