+++ api/src/main/java/shop/ink3/api/book/booktag/repository/BookTagRepository.java
@@ -1,9 +1,9 @@
-package shop.ink3.api.book.bookTag.repository;
+package shop.ink3.api.book.booktag.repository;
 
 import java.util.List;
 import org.springframework.data.jpa.repository.EntityGraph;
 import org.springframework.data.jpa.repository.JpaRepository;
-import shop.ink3.api.book.bookTag.entity.BookTag;
+import shop.ink3.api.book.booktag.entity.BookTag;
 
 public interface BookTagRepository extends JpaRepository<BookTag, Long> {
     @EntityGraph(attributePaths = "tag")
