+++ api/src/main/java/shop/ink3/api/coupon/coupon/service/Impl/CouponServiceImpl.java
@@ -15,9 +15,8 @@ import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import shop.ink3.api.book.book.entity.Book;
 import shop.ink3.api.book.book.repository.BookRepository;
-import shop.ink3.api.book.bookCategory.repository.BookCategoryRepository;
+import shop.ink3.api.book.bookcategory.repository.BookCategoryRepository;
 import shop.ink3.api.book.category.entity.Category;
-import shop.ink3.api.book.category.exception.CategoryNotFoundException;
 import shop.ink3.api.book.category.repository.CategoryRepository;
 import shop.ink3.api.book.category.service.CategoryService;
 import shop.ink3.api.common.dto.PageResponse;
