+++ api/src/main/java/shop/ink3/api/coupon/store/service/CouponStoreService.java
@@ -15,8 +15,7 @@ import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import shop.ink3.api.book.book.entity.Book;
 import shop.ink3.api.book.book.repository.BookRepository;
-import shop.ink3.api.book.bookCategory.entity.BookCategory;
-import shop.ink3.api.book.bookCategory.repository.BookCategoryRepository;
+import shop.ink3.api.book.bookcategory.repository.BookCategoryRepository;
 import shop.ink3.api.book.category.repository.CategoryRepository;
 import shop.ink3.api.book.category.service.CategoryService;
 import shop.ink3.api.coupon.bookCoupon.entity.BookCouponRepository;
